Added new classes Address and HomeType and all features for them. Fixed some bugs.

+++ ApplicationLayer/Interfaces/IService.cs
@@ -3,9 +3,9 @@
     public interface IService<TDTO> where TDTO : class
     {
         public List<TDTO> GetAll();
-        public TDTO GetById(long id);
+        public TDTO GetById(int id);
         public TDTO Save(TDTO dto);
-        public TDTO UpdateById(long id, TDTO dto);
-        public void DeleteById(long id);
+        public TDTO UpdateById(int id, TDTO dto);
+        public void DeleteById(int id);
     }
 }
