+++ DataAccessLayer/Interfaces/IRepository.cs
@@ -3,7 +3,7 @@
     public interface IRepository<TEntity> where TEntity : class
     {
         public List<TEntity> GetAll();
-        public TEntity Get(long id);
+        public TEntity Get(int id);
         public void Add(TEntity entity);
         public void Update(TEntity entity);
         public void Delete(TEntity entity);
